Work around google.oauth2 expectation of JSON in refresh error
It looks like this commit introduced an error where google.oauth2
library now expects valid JSON:
https://github.com/googleapis/google-auth-library-python/commit/eae1dcbbaba543b7d6dcab8fa40a322ff5b3d541#diff-1b020acb6247d690d2ca64fcb04c8f53

To work around it this just modifies the test to return valid JSON with
an error body.

Also I noticed the URL has changed, so this expands the httmock matcher
to respond for both APIs.